+++ pajbot/apiwrappers/twitch/helix.py
@@ -740,6 +740,25 @@ class TwitchHelixAPI(BaseTwitchAPI):
             },
         )
 
+    def update_follower_mode(
+        self,
+        channel_id: str,
+        bot_id: str,
+        authorization,
+        state: bool,
+        duration_m: Optional[int] = None,
+    ) -> None:
+        """Calls the _update_chat_settings function using the state and duration_m parameters.
+        duration_m is in minutes.
+        bot_id must match the user ID in authorization."""
+        self._update_chat_settings(
+            channel_id,
+            bot_id,
+            authorization,
+            follower_mode=state,
+            follower_mode_duration=duration_m,
+        )
+
     def update_emote_only_mode(self, channel_id: str, bot_id: str, authorization, emote_mode: bool):
         """Calls the _unique_chat_settings function using the emote_mode parameter.
         channel_id, bot_id and emote_mode are all required fields. bot_id must match the user ID in authorization."""